api_parser/requirements.txt
beautifulsoup4==4.11.1
bs4==0.0.1
celery==5.2.6
Django==3.2.13
django-celery-beat==2.2.1
django-celery-results==2.3.1
django-timezone-field==4.2.3
djangogrpcframework==0.2.1
grpcio==1.46.0
grpcio-tools==1.46.0
protobuf==3.20.1
djangorestframework==3.13.1
fake-useragent==0.1.11
flower==1.0.0
lxml==4.8.0
prompt-toolkit==3.0.29
psycopg2-binary==2.9.3
pyparsing==3.0.8
redis==4.2.2
requests==2.27.1
soupsieve==2.3.2.post1

